Extract API calls into custom hooks
